# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type GenerateQuestionsInputInterviewType = typeof GenerateQuestionsInputInterviewType[keyof typeof GenerateQuestionsInputInterviewType];


export const GenerateQuestionsInputInterviewType = {
  Technical_Interview: 'Technical Interview',
  HR_Interview: 'HR Interview',
  Behavioral_Interview: 'Behavioral Interview',
  General_Interview: 'General Interview',
} as const;

export type GenerateQuestionsInputDifficulty = typeof GenerateQuestionsInputDifficulty[keyof typeof GenerateQuestionsInputDifficulty];


export const GenerateQuestionsInputDifficulty = {
  Easy: 'Easy',
  Medium: 'Medium',
  Hard: 'Hard',
} as const;

export type GenerateQuestionsInputNumberOfQuestions = typeof GenerateQuestionsInputNumberOfQuestions[keyof typeof GenerateQuestionsInputNumberOfQuestions];


export const GenerateQuestionsInputNumberOfQuestions = {
  NUMBER_5: 5,
  NUMBER_10: 10,
} as const;

export interface GenerateQuestionsInput {
  interviewType: GenerateQuestionsInputInterviewType;
  /**
     * @minLength 2
     * @maxLength 80
     */
  role: string;
  difficulty: GenerateQuestionsInputDifficulty;
  numberOfQuestions: GenerateQuestionsInputNumberOfQuestions;
}

export interface InterviewQuestion {
  /** @minLength 8 */
  question: string;
}

export interface GeneratedQuestions {
  /** @minItems 1 */
  questions: InterviewQuestion[];
}

export interface AnsweredInterviewQuestion {
  /** @minLength 8 */
  question: string;
  /**
     * @minLength 1
     * @maxLength 8000
     */
  answer: string;
}

export type EvaluateInterviewInputInterviewType = typeof EvaluateInterviewInputInterviewType[keyof typeof EvaluateInterviewInputInterviewType];


export const EvaluateInterviewInputInterviewType = {
  Technical_Interview: 'Technical Interview',
  HR_Interview: 'HR Interview',
  Behavioral_Interview: 'Behavioral Interview',
  General_Interview: 'General Interview',
} as const;

export type EvaluateInterviewInputDifficulty = typeof EvaluateInterviewInputDifficulty[keyof typeof EvaluateInterviewInputDifficulty];


export const EvaluateInterviewInputDifficulty = {
  Easy: 'Easy',
  Medium: 'Medium',
  Hard: 'Hard',
} as const;

export interface EvaluateInterviewInput {
  interviewType: EvaluateInterviewInputInterviewType;
  /**
     * @minLength 2
     * @maxLength 80
     */
  role: string;
  difficulty: EvaluateInterviewInputDifficulty;
  /**
     * @minItems 1
     * @maxItems 10
     */
  questions: AnsweredInterviewQuestion[];
}

export interface EvaluatedInterviewQuestion {
  question: string;
  answer: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  score: number;
  feedback: string;
}

export interface InterviewEvaluation {
  /**
     * @minimum 0
     * @maximum 100
     */
  overallScore: number;
  performanceLevel: string;
  strengths: string[];
  improvements: string[];
  /**
     * @minItems 1
     * @maxItems 10
     */
  questions: EvaluatedInterviewQuestion[];
}