# lib/api-zod/src/generated/types/answeredInterviewQuestion.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AnsweredInterviewQuestion {
  /** @minLength 8 */
  question: string;
  /**
     * @minLength 1
     * @maxLength 8000
     */
  answer: string;
}