# lib/api-zod/src/generated/types/evaluateInterviewInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnsweredInterviewQuestion } from './answeredInterviewQuestion';
import type { EvaluateInterviewInputDifficulty } from './evaluateInterviewInputDifficulty';
import type { EvaluateInterviewInputInterviewType } from './evaluateInterviewInputInterviewType';

export interface EvaluateInterviewInput {
  interviewType: EvaluateInterviewInputInterviewType;
  /**
     * @minLength 2
     * @maxLength 80
     */
  role: string;
  difficulty: EvaluateInterviewInputDifficulty;
  /**
     * @minItems 1
     * @maxItems 10
     */
  questions: AnsweredInterviewQuestion[];
}