# lib/api-zod/src/generated/types/evaluatedInterviewQuestion.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EvaluatedInterviewQuestion {
  question: string;
  answer: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  score: number;
  feedback: string;
}