# lib/api-zod/src/generated/types/generateQuestionsInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GenerateQuestionsInputDifficulty } from './generateQuestionsInputDifficulty';
import type { GenerateQuestionsInputInterviewType } from './generateQuestionsInputInterviewType';
import type { GenerateQuestionsInputNumberOfQuestions } from './generateQuestionsInputNumberOfQuestions';

export interface GenerateQuestionsInput {
  interviewType: GenerateQuestionsInputInterviewType;
  /**
     * @minLength 2
     * @maxLength 80
     */
  role: string;
  difficulty: GenerateQuestionsInputDifficulty;
  numberOfQuestions: GenerateQuestionsInputNumberOfQuestions;
}