# lib/api-zod/src/generated/types/generatedQuestions.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InterviewQuestion } from './interviewQuestion';

export interface GeneratedQuestions {
  /** @minItems 1 */
  questions: InterviewQuestion[];
}