# lib/api-zod/src/generated/types/interviewEvaluation.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EvaluatedInterviewQuestion } from './evaluatedInterviewQuestion';

export interface InterviewEvaluation {
  /**
     * @minimum 0
     * @maximum 100
     */
  overallScore: number;
  performanceLevel: string;
  strengths: string[];
  improvements: string[];
  /**
     * @minItems 1
     * @maxItems 10
     */
  questions: EvaluatedInterviewQuestion[];
}